# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Dokumentasi API PraktikumApi",
      "description": "REVI WAHYU ABDILLAH | G211220039",
      "contact": {
        "email": "email@example.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/"
      }
    ],
    "paths": {
      "/api/mahasiswas": {
        "get": {
          "tags": [
            "Mahasiswa"
          ],
          "summary": "Dapatkan daftar semua mahasiswa",
          "description": "Mengembalikan daftar semua mahasiswa",
          "operationId": "getMahasiswa",
          "responses": {
            "200": {
              "description": "Daftar mahasiswa",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Mahasiswa"
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        },
        "post": {
          "tags": [
            "Mahasiswa"
          ],
          "summary": "Tambah mahasiswa baru",
          "description": "Membuat data mahasiswa baru",
          "operationId": "createMahasiswa",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mahasiswa"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Mahasiswa berhasil ditambahkan"
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      },
      "/api/mahasiswas/{nim}": {
        "get": {
          "tags": [
            "Mahasiswa"
          ],
          "summary": "Dapatkan detail mahasiswa",
          "description": "Mengembalikan detail mahasiswa berdasarkan NIM",
          "operationId": "getMahasiswaByNim",
          "parameters": [
            {
              "name": "nim",
              "in": "path",
              "required": true,
              "description": "NIM mahasiswa",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Detail mahasiswa",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Mahasiswa"
                  }
                }
              }
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        },
        "put": {
          "tags": [
            "Mahasiswa"
          ],
          "summary": "Perbarui data mahasiswa",
          "description": "Memperbarui data mahasiswa berdasarkan NIM",
          "operationId": "updateMahasiswa",
          "parameters": [
            {
              "name": "nim",
              "in": "path",
              "required": true,
              "description": "NIM mahasiswa",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mahasiswa"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Mahasiswa berhasil diperbarui"
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        },
        "delete": {
          "tags": [
            "Mahasiswa"
          ],
          "summary": "Hapus mahasiswa",
          "description": "Menghapus data mahasiswa berdasarkan NIM",
          "operationId": "deleteMahasiswa",
          "parameters": [
            {
              "name": "nim",
              "in": "path",
              "required": true,
              "description": "NIM mahasiswa",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Mahasiswa berhasil dihapus"
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      }
    },
    "components": {
      "schemas": {
        "Mahasiswa": {
          "title": "Mahasiswa",
          "description": "Model Mahasiswa",
          "properties": {
            "nim": {
              "description": "Nomor Induk Mahasiswa",
              "type": "string"
            },
            "nama": {
              "description": "Nama mahasiswa",
              "type": "string"
            },
            "alamat": {
              "description": "Alamat mahasiswa",
              "type": "string"
            },
            "tanggal_lahir": {
              "description": "Tanggal lahir mahasiswa",
              "type": "string",
              "format": "date"
            },
            "jurusan": {
              "description": "Jurusan mahasiswa",
              "type": "string"
            },
            "created_at": {
              "description": "Waktu pembuatan data",
              "type": "string",
              "format": "date-time"
            },
            "updated_at": {
              "description": "Waktu pembaruan data",
              "type": "string",
              "format": "date-time"
            }
          },
          "type": "object"
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "bearerFormat": "JWT",
          "scheme": "bearer"
        }
      }
    },
    "tags": [
      {
        "name": "Mahasiswa",
        "description": "Operasi terkait Mahasiswa"
      }
    ]
  }
  